requirements.txt
python-telegram-bot==20.8
Flask
nest_asyncio
aiofiles
psycopg2-binary